Make psychTestR test

The top-level function that defines a psychTestR test. This should be the final line of your app script, typically entitled app.R.

make_test(elts, opt = demo_options(), custom_admin_panel = NULL)

Arguments

elts

List of test elements defining the test's timeline.

opt

Options list as created by test_options().

custom_admin_panel

This argument supports the construction of a custom admin panel. The argument should be a function that runs during the Shiny server function. The argument list of this function should include ...; it may also include state, input, output, and session, all with their traditional meanings from Shiny or psychTestR. This function should include an expression of the form output$custom_admin_panel <- shiny::renderUI(...) where the code inside renderUI() defines the UI of the admin panel using standard Shiny UI vocabulary. Further documentation for this option should be forthcoming.