Deprecated functions in package psychTestR.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

loop_while(test, logic)

loop_while

We recommend that new projects use while_loop, which better resembles while loops in traditional programming languages.