Timeline

Timelines are series of psychTestR test elements that chain together to form a test. They support internationalisation, defining parallel series of test elements for the available languages.

timeline

Format

An object of class R6ClassGenerator of length 24.

Creation

Timelines are created using new_timeline.

Manipulation

Timelines can be combined with other timelines and with test elements using c.

Usage

Timelines are ultimately passed to make_test.

Other methods

x$get(language, i) returns a list of test elements corresponding to language as extracted from the timeline x, or, if i is not NULL, the ith such test element.

x$drop_languages(drop) removes support for a set of languages from timeline x, where drop is the character vector of languages to remove.