Trigger button

A version of shiny::actionButton that triggers the next psychTestR page.

trigger_button(
  inputId,
  label,
  icon = NULL,
  width = NULL,
  enable_after = 0,
  style = "",
  ...
)

Arguments

inputId
The input slot that will be used to access the value.

label
The contents of the button or link--usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon() to appear on the button.

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

enable_after
Number of seconds after which responses should be permitted.

style
CSS style information (character scalar).

...
Named attributes to be applied to the button or link.