Slider page

Creates a page where the participant responds by manipulating a slider.

slider_page(
  label,
  prompt,
  min,
  max,
  value,
  save_answer = TRUE,
  button_text = "Next",
  on_complete = NULL,
  admin_ui = NULL,
  step = NULL,
  round = FALSE,
  ticks = TRUE,
  animate = FALSE,
  width = NULL,
  sep = ",",
  pre = NULL,
  post = NULL,
  timeFormat = NULL,
  timezone = NULL,
  dragRange = TRUE
)

Arguments

label
Label for the current page (character scalar).

prompt
Prompt to display (character scalar or Shiny tag object).

min
The minimum value (inclusive) that can be selected.

max
The maximum value (inclusive) that can be selected.

value
The initial value of the slider. A numeric vector of length one will create a regular slider; a numeric vector of length two will create a double-ended range slider. A warning will be issued if the value doesn't fit between min and max.

save_answer
Whether or not to save the answer.

button_text
Text for the submit button (character scalar).

on_complete
Optional function to execute on leaving the page (after successful validation). The argument list should include ..., and any of: state, the participant's state object; answer, the participant's most recent answer; input, the current page's Shiny input object; session, the current Shiny session object; opt, the test's option list as created by test_options().

admin_ui
Optional UI component for the admin panel.

step
Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size). If the values are dates, step is in days; if the values are times (POSIXt), step is in seconds.

round
TRUE to round all values to the nearest integer; FALSE if no rounding is desired; or an integer to round to that number of digits (for example, 1 will round to the nearest 10, and -2 will round to the nearest .01). Any rounding will be applied after snapping to the nearest step.

ticks
FALSE to hide tick marks, TRUE to show them according to some simple heuristics.

animate
TRUE to show simple animation controls with default settings; FALSE not to; or a custom settings list, such as those created using animationOptions().

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

sep
Separator between thousands places in numbers.

pre
A prefix string to put in front of the value.

post
A suffix string to put after the value.

timeFormat
Only used if the values are Date or POSIXt objects. A time format string, to be passed to the Javascript strftime library. See https://github.com/samsonjs/strftime for more details. The allowed format specifications are very similar, but not identical, to those for R's base::strftime() function. For Dates, the default is "%F" (like "2015-07-01"), and for POSIXt, the default is "%F %T" (like "2015-07-01 15:32:10").

timezone
Only used if the values are POSIXt objects. A string specifying the time zone offset for the displayed times, in the format "+HHMM" or "-HHMM". If NULL (the default), times will be displayed in the browser's time zone. The value "+0000" will result in UTC time.

dragRange
This option is used only if it is a range slider (with two values). If TRUE (the default), the range can be dragged. In other words, the min and max can be dragged together. If FALSE, the range cannot be dragged.

Note

The RStudio preview function seems not to work for slider pages on some machines.