Timeline

Timelines are series of psychTestR test elements that chain together to form a test. They support internationalisation, defining parallel series of test elements for the available languages.

Creation

Timelines are created using new_timeline.

Manipulation

Timelines can be combined with other timelines and with test elements using c.

Usage

Timelines are ultimately passed to make_test.

Other methods

x$get(language, i) returns a list of test elements corresponding to language as extracted from the timeline x, or, if i is not NULL, the ith such test element.

x$drop_languages(drop) removes support for a set of languages from timeline x, where drop is the character vector of languages to remove.

Methods

Public methods

timeline$new()

timeline$get()

timeline$drop_languages()

timeline$print()

timeline$clone()

Method new()

Usage

timeline$new(x)

Method get()

Usage

timeline$get(language, i = NULL)

Method drop_languages()

Usage

timeline$drop_languages(drop)

Method print()

Usage

timeline$print(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

timeline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.